Making late command line options available for local runtime
